-Completed the remote, local and repository for book feature functions with some tests

+++ app/src/main/java/com/ayoprez/sobuu/shared/models/UserBookRating.kt
@@ -3,5 +3,5 @@ package com.ayoprez.sobuu.shared.models
 data class UserBookRating(val id: String,
                           val book: Book,
                           val user: Profile,
-                          val rating: Byte,
+                          val rating: Double,
                           val review: String)
